+++ src/decode.cpp
@@ -8,11 +8,11 @@ std::vector<Data> decode(const uint8_t* image, size_t width, size_t height) {
     Decoder decoder;
     decoder.fill_image(image, width, height);
 
-    auto num_codes = decoder.count();
+    auto num_codes = static_cast<size_t>(decoder.count());
     std::vector<Data> parsed_codes;
     parsed_codes.reserve(num_codes);
 
-    for (int i=0; i < num_codes; i++) {
+    for (size_t i=0; i < num_codes; i++) {
         try {
             parsed_codes.push_back(decoder.decode_index(i));
         } catch(std::runtime_error&) {
